Rename jamsite-pages-dev to jamsite-dev
