+++ lib/index.js
@@ -1,6 +1,6 @@
 const { micro, getRequestHandler } = require('@jamsite/jamsite')
 const bs = require('browser-sync').create()
-const JamsitePagesDev = require('./jamsite-pages-dev')
+const JamsitePagesDev = require('./jamsite-dev')
 
 module.exports.start = function start () {
   const config = require('../config')
