+++ lib/jamsite-dev.js
@@ -1,9 +1,9 @@
 const path = require('path')
-const { JamsitePages } = require('@jamsite/jamsite')
+const { Jamsite } = require('@jamsite/jamsite')
 
 const EVENT_UPDATE = 'update'
 
-class JamsitePagesDev extends JamsitePages {
+class JamsiteDev extends Jamsite {
   defaultConfig () {
     const config = super.defaultConfig()
     config.chokidarOptions.persistent = true
@@ -30,6 +30,6 @@ class JamsitePagesDev extends JamsitePages {
   }
 }
 
-JamsitePagesDev.EVENT_UPDATE = EVENT_UPDATE
+JamsiteDev.EVENT_UPDATE = EVENT_UPDATE
 
-module.exports = JamsitePagesDev
+module.exports = JamsiteDev
